+++ template/vue-mfe-sub/src/index.js
@@ -1,5 +1,5 @@
 import Vue from "vue";
-import App from "./app.vue";
+import App from "./App.vue";
 
 const root = document.createElement("div");
 document.body.appendChild(root);